remove all references to SelectedAccount component
