+++ src/views/Wallet/views/Account/Send/Container.js
@@ -1,45 +1,38 @@
 /* @flow */
 
-
-import * as React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import { reconnect } from 'actions/DiscoveryActions';
 import SendFormActions from 'actions/SendFormActions';
 import type { MapStateToProps, MapDispatchToProps } from 'react-redux';
 import type { State, Dispatch } from 'flowtype';
-import type { StateProps as BaseStateProps, DispatchProps as BaseDispatchProps } from 'views/Wallet/components/SelectedAccount';
 import AccountSend from './index';
 
 type OwnProps = { }
 
-export type StateProps = BaseStateProps & {
+export type StateProps = {
+    selectedAccount: $ElementType<State, 'selectedAccount'>,
     sendForm: $ElementType<State, 'sendForm'>,
+    wallet: $ElementType<State, 'wallet'>,
     fiat: $ElementType<State, 'fiat'>,
     localStorage: $ElementType<State, 'localStorage'>,
-    children?: React.Node;
 }
 
-export type DispatchProps = BaseDispatchProps & {
+export type DispatchProps = {
     sendFormActions: typeof SendFormActions,
 }
 
-export type Props = StateProps & BaseStateProps & DispatchProps & BaseDispatchProps;
+export type Props = StateProps & DispatchProps;
 
 const mapStateToProps: MapStateToProps<State, OwnProps, StateProps> = (state: State): StateProps => ({
-    className: 'send-from',
     selectedAccount: state.selectedAccount,
-    wallet: state.wallet,
-    blockchain: state.blockchain,
-
     sendForm: state.sendForm,
+    wallet: state.wallet,
     fiat: state.fiat,
     localStorage: state.localStorage,
 });
 
 const mapDispatchToProps: MapDispatchToProps<Dispatch, OwnProps, DispatchProps> = (dispatch: Dispatch): DispatchProps => ({
-    blockchainReconnect: bindActionCreators(reconnect, dispatch),
     sendFormActions: bindActionCreators(SendFormActions, dispatch),
 });
